services/user_service/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv